+++ riverdriver/riverpgxv5/migration/main/007_river_leader_non_default_name.up.sql
@@ -1,0 +1,3 @@
+ALTER TABLE /* TEMPLATE: schema */river_leader
+    DROP CONSTRAINT name_length,
+    ADD CONSTRAINT name_length CHECK (char_length(name) > 0 AND char_length(name) < 128);