+++ rivershared/testfactory/test_factory.go
@@ -111,6 +111,7 @@ type LeaderOpts struct {
 	ElectedAt *time.Time
 	ExpiresAt *time.Time
 	LeaderID  *string
+	Name      *string
 	Now       *time.Time
 	Schema    string
 }
@@ -122,6 +123,7 @@ func Leader(ctx context.Context, tb testing.TB, exec riverdriver.Executor, opts 
 		ElectedAt: opts.ElectedAt,
 		ExpiresAt: opts.ExpiresAt,
 		LeaderID:  ptrutil.ValOrDefault(opts.LeaderID, "test-client-id"),
+		Name:      ptrutil.ValOrDefault(opts.Name, "default"),
 		Now:       opts.Now,
 		Schema:    opts.Schema,
 		TTL:       10 * time.Second,